fix: GISScalar query variables
There was an issue when using query variables and WKT
input. While parsing the input value, it would look for
node.value which is null for a string.

Remedied that. On another note, these two lines were not
being covered by any test case and were reported as line
misses by pytest-cov and that was the root of all evil.
We're still missing code-cov by 2 lines, but that's a story
for another day.

Thanks to Dimitri Gnidash.

Fixes: https://github.com/EverWinter23/graphene-gis/issues/9